auth && get blogs by category implemented && views added

+++ BlogAppCore/Core/IRepositories/IBlogRepository.cs
@@ -4,7 +4,7 @@ namespace BlogAppCore.Core.IRepositories
 {
     public interface IBlogRepository : IGenericRepository<Blog>
     {
-        public Task<IEnumerable<Blog>> GetByCategoryIdAsync(int id);
+        public Task<IEnumerable<Blog>> GetByCategoryIdAsync(long id);
         public Task<IEnumerable<Blog>> GetByUserIdAsync(string id);
         public Task<IEnumerable<Blog>> GetDeletedBlogsByUserIdAsync(string id);
     }
